New config option `testbuildonly`
When this option is specified, all targets with a name matching
`*_test` only get their dependencies built, but are not linked.

This is intended to validate source code validity for tests without
incurring the cost of linking.

This patch uses a new option (as opposed to overloading `linkmode`)
because it needs to affect only the `_test` targets; not all targets.